requirements.txt
aiohttp
python-telegram-bot
pandas
numpy
matplotlib
scikit-learn
backtrader
pytest